Refactor localisation initialisation
Localisation is now initialised in AmuletApp and upon first usage for resource packs.
Remove global Signal
