+++ src/amulet/app/app/_app.py
@@ -1,13 +1,10 @@
-from PySide6.QtCore import QObject, Signal
+from PySide6.QtWidgets import QApplication
 
 from amulet.app.localisation._app import init as init_localisation
 
 
-class AppCreatedObject(QObject):
-    app_created = Signal()
+class AmuletApp(QApplication):
+    def __init__(self, *args, **kwargs) -> None:
+        super().__init__(*args, **kwargs)
 
-
-_obj = AppCreatedObject()
-app_created = _obj.app_created
-
-app_created.connect(init_localisation)
+        init_localisation()
