+++ src/plugin/amulet/resource_pack/__init__.py
@@ -1,3 +1,1 @@
-from . import _localisation
-
 from ._api import get_resource_pack_handle
